refactor: Refactor confirmation dialogs
Create a one general method showConfirmationDialog in the class
DialogUtils that shows a title and a message given as parameters
by callers methods from other classes.

+++ src/main/java/pl/magdalena/brejna/colourtheworldapp/models/ProjectList.java
@@ -22,7 +22,7 @@ public class ProjectList {
     }
 
     private void showDeleteConfirmationDialog(Project project){
-        DialogsUtils.showDeleteConfirmationDialog()
+        DialogsUtils.showConfirmationDialog("delete.title", "delete.text")
                 .filter(response -> response == ButtonType.OK)
                 .ifPresent(response -> deleteProject(project));
     }
